flask/requirements.txt
anyio==3.7.0
asgiref==3.7.2
blinker==1.6.2
certifi==2023.5.7
click==8.1.4
colorama==0.4.6
distlib==0.3.6
Django==4.2.2
django-cors-headers==4.0.0
djangorestframework==3.14.0
dnspython==2.3.0
exceptiongroup==1.1.2
filelock==3.12.2
Flask==2.3.2
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.24.3
platformdirs==3.8.1
pymongo==4.3.3
python-telegram-bot==20.3
pytz==2023.3
sniffio==1.3.0
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
virtualenv==20.23.1
Werkzeug==2.3.6
